Should improve the odds of catching common fish, even for weak fishing
Even if you have no skill level in fishing, the chance to catch a fish should never fall below the odds assigned by your bait.

Since you do not gain experience unless you catch fish and the probability is rolled twice, you have two chances to not catch a fish. This aims to soften the blow of that.