Adds new cleric/acolyte of necra spells
Adds a tier 3 (500 devotion) touch spell for acolytes and Clerics of Necra that has a moderate amount of force and does toxin (necrotic) damage.

Can only be used once a minute.

In the future I am looking into restoring speak with dead for ONLY the acolyte of Necra so they can be the death detective and have more of a niche.

Looking into t3 spells for dendor clerics/acolytes, who are also currently lacking one.