Removes (most) Drifter Class Rolls
Drifter class rolls and quantity limitations (for the most part) are antithetical toward roleplay. They favour a game-y attitude of changing the character you've selected to suit the cards you're given, rather than the reverse.

This removes all of the class rolls except for assassin due to its antagonistic nature. It also removes the quantity limits on all non-magical and non-nobility classes.

I don't doubt that people will find issues quickly with this, but this is something that should be fixed through class balance, not an arbitrary padlock that sometimes falls off the door.