Elder Skill Additions and Adjustements
Gives the elder riding skill, as the new Dakka map gives him a saiga to ride round-start, much like the king. Very useful for both RP and him getting around as the top dog from the village to the town.

Also gives wrestling, because why didn't he have it as is? He already has no armour training or dodge expertise, letting him be grab bait is horrible. Did lower the fistfighting from 4 to 3, though, to make up for the 3 in wrestling. Puts him on par with other classes with combat skills.

Athletics is just a tiny bonus.