Village Elder > Mayor
Guayo wanted the Village Elder to be turned into the mayor, so that's why the courthouse was given to him within the town.

This just adjusts his title, some spawners and the like, gives him a keyring and removes the extra keys from his house.

Also strips him of his cloak, cudgel and replaces his shirt with one of the new tunics.