also slightly lowers athletics on poacher
slightly lowers it, so they have skilled athletics instead of expert. This is to make the brigand have an extra very small edge over the others, since I feel like while they get a lot of skills, they miss out on a lot of things the other classes get as well.