looking at the wanted poster actually works
if you stare at a wanted poster, you will actually be able to see if someone is a bandit. When examining someone you'll be notified if they are a bandit, but only if you had looked at a wanted poster at some point previously.

Staring at a wanted poster as a bandit will just notify you that yup, your face is still on there.