Village Public Mine
There was this giant pile of randomized ore rock that was completely inaccessible, so I decided to do something with it. Now there's a small mine there for the public to use (nowhere near as good as the actual one), with a small little surprise attached to a piece of it for anyone brave enough to go poke their head in there.