Destroyed Grenzelhoft Camp
An outpost made by what appears to be a scouting party of the Black Empire. Why exactly they were there, who knows, but it seems that Graggar's spawn have destroyed it. The loot is VERY good if you can get it, but good luck. You're going to need friends for this.