Adds a small southern dungeon section to the Raytown sewers, as well as a small mine in the Village (#873)
* Update roguetown.dmm

* Update roguetown.dmm

* Sewer Dungeon Expansion

Adds a neat skrallax cave to the southern most part of the underground map, as well as makes a secret alchemy lair adjacent. Both can be accessed by the sewers, and both are quite dangerous to compensate for the rewards.

* Village Public Mine

There was this giant pile of randomized ore rock that was completely inaccessible, so I decided to do something with it. Now there's a small mine there for the public to use (nowhere near as good as the actual one), with a small little surprise attached to a piece of it for anyone brave enough to go poke their head in there.

* Update roguetown.dmm

* Update roguetown.dmm

* Update roguetown.dmm

* Update roguetown.dmm

---------

Co-authored-by: Optimism333 <81007356+Optimism333@users.noreply.github.com>